inline master.log.test.js
