+++ src/master/master.test.js
@@ -11,11 +11,10 @@ import * as ActionCreators from '../core/action-creators';
 import * as Redux from 'redux';
 import { InMemory } from '../server/db/inmemory';
 
-// silence log output during these tests
-console.log = () => {};
+console.log = jest.fn();
 const { GameMaster } = require('./master');
 
-const game = Game({ seed: 0 });
+const game = Game({ seed: 0, flow: { setActionPlayers: true } });
 
 function TransportAPI(send = jest.fn(), sendAll = jest.fn()) {
   return { send, sendAll };
@@ -190,22 +189,44 @@ describe('update', async () => {
   test('invalid gameID', async () => {
     await master.onUpdate(action, 1, 'unknown', '1');
     expect(sendAll).not.toHaveBeenCalled();
+    expect(console.log).toHaveBeenCalledWith(
+      `ERROR: game not found, gameID=[unknown]`
+    );
   });
 
   test('invalid stateID', async () => {
     await master.onUpdate(action, 100, 'gameID', '1');
     expect(sendAll).not.toHaveBeenCalled();
+    expect(console.log).toHaveBeenCalledWith(
+      `ERROR: invalid stateID, was=[100], expected=[1]`
+    );
   });
 
   test('invalid playerID', async () => {
     await master.onUpdate(action, 1, 'gameID', '100');
     await master.onUpdate(ActionCreators.makeMove(), 1, 'gameID', '100');
     expect(sendAll).not.toHaveBeenCalled();
+    expect(console.log).toHaveBeenCalledWith(
+      `ERROR: event not processed - invalid playerID=[100]`
+    );
   });
 
   test('valid gameID / stateID / playerID', async () => {
     await master.onUpdate(action, 1, 'gameID', '1');
     expect(sendAll).toHaveBeenCalled();
+  });
+
+  test('writes log when player is not an action player', async () => {
+    const setActionPlayersEvent = ActionCreators.gameEvent('setActionPlayers', [
+      '1',
+    ]);
+    await master.onUpdate(setActionPlayersEvent, 2, 'gameID', '0');
+
+    const move = ActionCreators.makeMove('move');
+    await master.onUpdate(move, 3, 'gameID', '0');
+    expect(console.log).toHaveBeenCalledWith(
+      `ERROR: move not processed - canPlayerMakeMove=false, playerID=[0]`
+    );
   });
 });
